+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Capabilities for setting up alternate exchange rate source. Refs UIF-583.
 * Do not fail rollover results export when no rollover budgets are created. Refs UIF-607.
 * Apply CSS classes to the transaction modal content to ensure a minimal height. UIF-610.
+* Add donor information accordion to the fund view. Refs UIF-569.
 
 ## [8.0.4](https://github.com/folio-org/ui-finance/tree/v8.0.4) (2025-05-06)
 [Full Changelog](https://github.com/folio-org/ui-finance/compare/v8.0.3...v8.0.4)